backend/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
gast==0.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h5py==3.12.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.6.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.2
rsa==4.9
six==1.16.0
sqlite-vec==0.1.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tqdm==4.66.6
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
